+++ toys/CHANGELOG.md
@@ -12,6 +12,7 @@
 * CHANGED: Flags are no longer automatically considered to have a value if they have a default or acceptor but also have explicit flag syntax that does not include a value.
 * CHANGED: Exec reports failure to start processes in the result object rather than, e.g. raising ENOENT.
 * IMPROVED: Toys no longer displays a stack trace if a tool is interrupted.
+* IMPROVED: Error messages for flag groups are more complete.
 * FIXED: Prevented toys-core from being ousted from the load path if a toys file invoked bundler setup.
 * FIXED: Acceptors no longer raise errors when run on missing optional values.
 * FIXED: When reporting errors in toys files, the line number was off by 2.